+++ server/conn_test.go
@@ -679,93 +679,3 @@ func (ts *ConnTestSuite) TestPrefetchPointKeys(c *C) {
 	tk.MustExec("commit")
 	tk.MustQuery("select * from prefetch").Check(testkit.Rows("1 1 3", "2 2 6", "3 3 5"))
 }
-
-type ConnSerialTestSuite struct {
-	dom   *domain.Domain
-	store kv.Storage
-}
-
-var _ = SerialSuites(&ConnSerialTestSuite{})
-
-func (ts *ConnSerialTestSuite) SetUpSuite(c *C) {
-	testleak.BeforeTest()
-	var err error
-	ts.store, err = mockstore.NewMockStore()
-	c.Assert(err, IsNil)
-	ts.dom, err = session.BootstrapSession(ts.store)
-	c.Assert(err, IsNil)
-}
-
-func (ts *ConnSerialTestSuite) TearDownSuite(c *C) {
-	ts.dom.Close()
-	ts.store.Close()
-	testleak.AfterTest(c)()
-}
-
-func (ts *ConnSerialTestSuite) TestHandleQueryCleanUp(c *C) {
-	store, err := mockstore.NewMockStore()
-	c.Assert(err, IsNil)
-	defer store.Close()
-	dom, err := session.BootstrapSession(store)
-	c.Assert(err, IsNil)
-	defer dom.Close()
-	se, err := session.CreateSession4Test(store)
-	c.Assert(err, IsNil)
-	tc := &TiDBContext{
-		Session: se,
-		stmts:   make(map[int]*TiDBStatement),
-	}
-	_, err = se.Execute(context.Background(), "create table test.t(id int)")
-	c.Assert(err, IsNil)
-	var outBuffer bytes.Buffer
-	tidbdrv := NewTiDBDriver(ts.store)
-	cfg := newTestConfig()
-	cfg.Port, cfg.Status.StatusPort = 0, 0
-	cfg.Status.ReportStatus = false
-	server, err := NewServer(cfg, tidbdrv)
-	c.Assert(err, IsNil)
-	defer server.Close()
-	cc := &clientConn{
-		connectionID: 1,
-		salt:         []byte{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F, 0x10, 0x11, 0x12, 0x13, 0x14},
-		server:       server,
-		pkt: &packetIO{
-			bufWriter: bufio.NewWriter(&outBuffer),
-		},
-		collation:  mysql.DefaultCollationID,
-		peerHost:   "localhost",
-		alloc:      arena.NewAllocator(512),
-		ctx:        tc,
-		capability: 0,
-	}
-
-	originMaxConsumed := executor.GlobalMemoryUsageTracker.MaxConsumed()
-	defer executor.GlobalMemoryUsageTracker.SetMaxConsumed(originMaxConsumed)
-	executor.GlobalMemoryUsageTracker.SetMaxConsumed(0)
-	originConsume := executor.GlobalMemoryUsageTracker.BytesConsumed()
-
-	testcases := []struct {
-		sql string
-	}{
-		{
-			sql: "insert into test.t values (1)",
-		},
-		{
-			sql: "update test.t set id = 2 where id = 1",
-		},
-		{
-			sql: "select * from test.t",
-		},
-	}
-
-	for _, testcase := range testcases {
-		err := cc.handleQuery(context.Background(), testcase.sql)
-		c.Assert(err, IsNil)
-		afterConsume := executor.GlobalMemoryUsageTracker.BytesConsumed()
-		c.Assert(originConsume, Equals, afterConsume)
-	}
-
-	// assert consume happened
-	afterMaxConsumed := executor.GlobalMemoryUsageTracker.MaxConsumed()
-	c.Assert(afterMaxConsumed, Greater, int64(0))
-}
